more kernel refactoring

+++ src/kernel/syscalls/Crash.cpp
@@ -1,10 +1,10 @@
 #include <syscalls/Crash.hpp>
 
 Result<size_t> Crash(const shared_ptr<Process> &,
-            uint64_t,
-            uint64_t,
-            uint64_t,
-            uint64_t,
-            uint64_t) {
+                     size_t,
+                     size_t,
+                     size_t,
+                     size_t,
+                     size_t) {
     return Status::BadUserSpace();
 }
