+++ src/kernel/syscalls/Crash.hpp
@@ -4,8 +4,8 @@
 #include <processes/Process.hpp>
 
 Result<size_t> Crash(const shared_ptr<Process> &process,
-            uint64_t,
-            uint64_t,
-            uint64_t,
-            uint64_t,
-            uint64_t);
+                     size_t,
+                     size_t,
+                     size_t,
+                     size_t,
+                     size_t);
